chore(greenhouse): fixes use of authProvider
