+++ apps/greenhouse/src/hooks/useApi.js
@@ -5,12 +5,13 @@
 
 import { useCallback, useMemo } from "react"
 import { createClient } from "sapcc-k8sclient"
-import { useAuthData, useGlobalsApiEndpoint, useGlobalsAssetsHost } from "../components/StoreProvider"
+import { useGlobalsApiEndpoint, useGlobalsAssetsHost } from "../components/StoreProvider"
 import { createPluginConfig } from "../lib/plugin"
+import { useAuth } from "../components/AuthProvider"
 
 // get plugin configs from k8s api
 const useApi = () => {
-  const authData = useAuthData()
+  const { data: authData } = useAuth()
   const apiEndpoint = useGlobalsApiEndpoint()
   const assetsHost = useGlobalsAssetsHost()
 
@@ -27,6 +28,8 @@ const useApi = () => {
   }, [apiEndpoint, authData?.JWT])
 
   const getPluginConfigs = useCallback(() => {
+    console.info(">>>>>>>>>>>>>>>>>>>>>Fetching plugin configs", client, namespace)
+
     if (!client || !namespace) return Promise.resolve({})
 
     return client
